Update continue_training feature

+++ Reporter.py
@@ -1,3 +1,4 @@
+from torch._C import set_flush_denormal
 from path import Path
 import csv
 
@@ -8,9 +9,10 @@ class Reporter():
         print('=> will save everything to {}'.format(self.args.save_path))
         self.args.save_path.makedirs_p()
 
-        with open(self.args.save_path/self.args.log_summary, 'w') as csvfile:
-            writer = csv.writer(csvfile, delimiter='\t')
-            writer.writerow(['train_loss', 'validation_loss', 'abs_diff', 'abs_rel', 'sq_rel', 'a1', 'a2', 'a3'])
+        if not self.args.resume and (self.args.save_path/self.args.log_summary).exists():
+            with open(self.args.save_path/self.args.log_summary, 'w') as csvfile:
+                writer = csv.writer(csvfile, delimiter='\t')
+                writer.writerow(['train_loss', 'validation_loss', 'abs_diff', 'abs_rel', 'sq_rel', 'a1', 'a2', 'a3'])
 
     def update_log_summary(self, train_loss, decisive_error, errors):
         with open(self.args.save_path/self.args.log_summary, 'a') as csvfile:
